Don't remove packagesFolder if it already exists (#3910)
If compilefolders are used the packages folder will be deleted and
recreated during the pipeline run.

In certain cases especially when you use customized AL:Go with foreign
container systems, this folder exists on purpose and already includes
files. This files should not be deleted because it will lead the
pipeline to fail

+++ AppHandling/Run-AlPipeline.ps1
@@ -875,10 +875,12 @@ if ($updateLaunchJson) {
 
 if ($useCompilerFolder -or $filesOnly -or !$useDevEndpoint) {
     $packagesFolder = CheckRelativePath -baseFolder $baseFolder -sharedFolder $sharedFolder -path $packagesFolder -name "packagesFolder"
-    if (Test-Path $packagesFolder) {
-        Remove-Item $packagesFolder -Recurse -Force
+    if (!($bcContainerHelperConfig.doNotRemovePackagesFolderIfExists)) {
+        if (Test-Path $packagesFolder) {
+            Remove-Item $packagesFolder -Recurse -Force
+        }
     }
-    New-Item $packagesFolder -ItemType Directory | Out-Null
+    New-Item $packagesFolder -ItemType Directory -Force | Out-Null 
 }
 
 if ($useDevEndpoint) {
